Integrate the abstract classes into one and fix the Pcf8574 driver
