+++ src/main/java/com/pi4j/drivers/io/expander/pcf8574/Pcf8574InputDriver.java
@@ -1,36 +1,19 @@
 package com.pi4j.drivers.io.expander.pcf8574;
 
-import com.pi4j.drivers.io.expander.AbstractInputExpander;
 import com.pi4j.io.ListenableOnOffRead;
 import com.pi4j.io.i2c.I2C;
 
 /**
- * Input driver for the Pcf8574.
- * <p>
- * As the input and output functionality of this chip uses separate addresses, it seemed most straightforward
- * to implement these as separate classes.
+ * @deprecated Please use Pcf8574Driver instead.
  */
-public class Pcf8574InputDriver extends AbstractInputExpander {
-    private final I2C i2c;
+@Deprecated
+public class Pcf8574InputDriver extends Pcf8574Driver {
 
     /**
      * Creates a new PCF 8574 input driver. The interrupt pin will be used to trigger update
      * requests from the chip. If null, state changes can still be observed via the poll() method.
      */
     public Pcf8574InputDriver(I2C i2c, ListenableOnOffRead<?> interruptPin) {
-        super (8, interruptPin);
-        this.i2c = i2c;
-    }
-
-    @Override
-    protected int readInputsImpl() {
-        return i2c.read();
-
-    }
-
-    @Override
-    public void close() {
-        super.close();
-        i2c.close();
+        super (i2c, interruptPin, 0xff, 0xff);
     }
 }
