--- src/test/java/com/pi4j/examples/interrupts/JoystickExample.java
@@ -1,4 +1,0 @@
-package com.pi4j.examples.interrupts;
-
-public class JoystickExample {
-}
